% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{create_dot_forest_plot}
\alias{create_dot_forest_plot}
\title{Dot and Forest plots}
\usage{
create_dot_forest_plot(
  data,
  drug,
  benefit,
  risk,
  filters,
  category,
  type_graph,
  type_risk,
  select_nnx,
  x_scale_fixed_free,
  ci_method,
  space_btwn_out_yn = "Y"
)
}
\arguments{
\item{data}{(\code{data.frame}) dataset}

\item{drug}{(\code{character}) selected drug}

\item{benefit}{(\code{character}) selected benefit}

\item{risk}{(\code{character}) selected risk}

\item{filters}{(\code{character}) selected filter}

\item{category}{(\code{character}) selected category}

\item{type_graph}{(\code{character}) selected way to display binary outcomes}

\item{type_risk}{(\code{character}) selected way to display risk outcomes
(crude proportions, exposure-adjusted rates (per 100 PYs))}

\item{select_nnx}{(\code{character}) show NNT/NNH}

\item{x_scale_fixed_free}{(\code{character}) free or fixed x-axis scale}

\item{ci_method}{(\code{character}) selected method to display}

\item{space_btwn_out_yn}{(\code{character}) control spacing between outcomes
confidence intervals
(Supplied in effects table, Calculated within the program)}
}
\description{
Create Dot and Forest plots and associated data
}
\examples{
dot_plot_src <- subset(effects_table, !is.na(Prop1))
bdin <- subset(dot_plot_src, Factor == "Benefit")
rdin <- subset(dot_plot_src, Factor == "Risk")

create_dot_forest_plot(
  data = dot_plot_src,
  drug = unique(dot_plot_src$Trt1),
  benefit = unique(bdin$Outcome),
  risk = unique(rdin$Outcome),
  filters = "None",
  category = "All",
  type_graph = "Absolute risk",
  type_risk = "Crude proportions",
  select_nnx = "Y",
  x_scale_fixed_free = "Fixed",
  ci_method = "Calculated",
  space_btwn_out_yn = "N"
)
}
