% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_label_der.R
\name{create_order_label_der}
\alias{create_order_label_der}
\title{Dot Plot order label}
\usage{
create_order_label_der(
  indata,
  groupvars,
  groupeff,
  space_btwn_out_yn = "Y",
  dataout = NULL
)
}
\arguments{
\item{indata}{(\code{data.frame}) data frame for dot plot}

\item{groupvars}{variables to order by}

\item{groupeff}{effects table outcome order}

\item{space_btwn_out_yn}{control spacing between outcomes}

\item{dataout}{(\code{data.frame}) ordered data frame for dot plot}
}
\value{
data frame for specified type of analysis
}
\description{
Order dot plot labels and only show first occurrence of label
}
\examples{
dot_plot_src <- subset(effects_table, Type == "Binary")
dot_data <- prepare_dot_data_b(dot_plot_src)

dot_plot_data <- create_order_label_der(
  indata = dot_data,
  groupvars = c("type", "factor", "outcome"),
  groupeff = c("group"),
  space_btwn_out_yn = "N"
)

head(dot_plot_data)
}
