% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bands}
\alias{data_bands}
\title{Example bands data used for Figure 3}
\format{
\subsection{\code{data_bands}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{level}{level each band represents}
\item{ystart}{starting point of each band}
\item{yend}{ending point of each band}
\item{col}{color of each band}

}
}
}
\usage{
data_bands
}
\description{
Example bands data used for Figure 3
}
\keyword{datasets}
