% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_lines}
\alias{data_lines}
\title{\enumerate{
\item stage: stage each line represents
\item xstart: x-axis starting point of each line
\item xend: x-axis ending point of each line
\item y: y-axis position of each line
\item col: color of each line
\item xpos: x-axis position of each annotation (relative to xstart)
\item ypos: y-axis position of each annotation (relative to y)
}}
\format{
\subsection{\code{data_lines}}{

A data frame with 7 rows and 7 columns:
\describe{
\item{stage}{stage each line represents}
\item{xstart}{x-axis starting point of each line}
\item{xend}{x-axis ending point of each line}
\item{y}{y-axis position of each line}
\item{col}{color of each band}
\item{xpos}{x-axis position of each annotation (relative to xstart)}
\item{ypos}{y-axis position of each annotation (relative to y)}

}
}
}
\usage{
data_lines
}
\description{
Example lines data used for Figure 3
}
\keyword{datasets}
