% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{generate_fig_rft}
\alias{generate_fig_rft}
\title{Forest plot}
\usage{
generate_fig_rft(
  data,
  xlabel = NULL,
  ylabel = NULL,
  fact_subset = "Benefit",
  type_subset = "Binary",
  type_scale = "Fixed",
  scale_x = "Identity",
  x_scale_n1_p1 = "N",
  select_nnx
)
}
\arguments{
\item{data}{(\code{character}) Data for figure}

\item{xlabel}{(\code{character}) Label for x-axis}

\item{ylabel}{(\code{character}) Label for y-axis}

\item{fact_subset}{(\code{character}) filter data}

\item{type_subset}{(\code{character}) Selected subset "Binary" or "Continuous"}

\item{type_scale}{(\code{character}) selected scale display type}

\item{scale_x}{(\code{character}) Type of scale for x-axis "Identity" or
"log10"}

\item{x_scale_n1_p1}{(\code{character}) fix x-axis scale between -1 and 1}

\item{select_nnx}{(\code{character}) show NNT/NNH}
}
\description{
Create Forest plot
}
\examples{
dot_plot_src <- subset(effects_table, !is.na(Prop1))
bdin <- subset(dot_plot_src, Factor == "Benefit")
rdin <- subset(dot_plot_src, Factor == "Risk")

fplot_data <- prepare_dot_forest_plot_data(
  data = dot_plot_src,
  drug = unique(dot_plot_src$Trt1),
  benefit = unique(bdin$Outcome),
  risk = unique(rdin$Outcome),
  filters = "None",
  category = "All",
  type_graph = "Absolute risk",
  type_risk = "Crude proportions",
  ci_method = "Calculated"
)
forest_plot_data <- subset(
  fplot_data$forest_plot_data,
  factor == "Benefit" & type == "Binary"
)
generate_fig_rft(
  data = forest_plot_data,
  fact_subset = "Benefit",
  type_subset = "Binary",
  xlabel = "Treatment Difference with 95\% CI\n",
  select_nnx = "No"
)
}
