% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{populated_effects_table}
\alias{populated_effects_table}
\title{Populated effects table}
\usage{
populated_effects_table(data)
}
\arguments{
\item{data}{(\code{data.frame}) dataset}
}
\value{
Effects tble with calculated CIs
}
\description{
Generate populated effects table with calculated CIs
}
\examples{
effects_table <- populated_effects_table(effects_table)
head(effects_table)
}
