% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{prepare_absolute_risk_data}
\alias{prepare_absolute_risk_data}
\title{Forest plot data for absolute risk}
\usage{
prepare_absolute_risk_data(
  data,
  change_ref = c("Y", "N"),
  ci_method = c("Supplied", "Calculated")
)
}
\arguments{
\item{data}{(\code{data.frame}) Effects table}

\item{change_ref}{Change reference group ("Y" or "N)}

\item{ci_method}{(\code{character}) confidence interval method,
either \code{Supplied} (taken from the effect table)
or \code{Calculated} (calculated within the program).}
}
\value{
data frame for specified type of analysis
}
\description{
Prepare Forest plot data for absolute risk
}
\examples{
forest_plot_src <- subset(effects_table, !is.na(Prop1))
forest_plot_data <- prepare_absolute_risk_data(
  forest_plot_src,
  "Y",
  "Calculated"
)
head(forest_plot_data)
}
