% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{prepare_dot_data_b}
\alias{prepare_dot_data_b}
\title{Dot plot data for binary outcomes}
\usage{
prepare_dot_data_b(data)
}
\arguments{
\item{data}{(\code{data.frame}) Effects table with Binary outcomes}
}
\value{
data frame for specified type of analysis
}
\description{
Prepare Dot plot data for binary outcomes
}
\examples{
dot_plot_src <- subset(effects_table, Type == "Binary")
dot_plot_data <- prepare_dot_data_b(dot_plot_src)
head(dot_plot_data)
}
