% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{prepare_dot_data_exp_adj}
\alias{prepare_dot_data_exp_adj}
\title{Dot plot data for exposure-adjusted rates}
\usage{
prepare_dot_data_exp_adj(data)
}
\arguments{
\item{data}{(\code{data.frame}) Effects table}
}
\value{
data frame for specified type of analysis
}
\description{
Prepare Dot plot data for exposure-adjusted rates
}
\examples{
dot_plot_src <- subset(effects_table, !is.na(Rate_Type))
dot_plot_data <- prepare_dot_data_exp_adj(dot_plot_src)
head(dot_plot_data)
}
