% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_dot_plot_utils.R
\name{pre_proc}
\alias{pre_proc}
\alias{prepare_dot_forest_plot_data}
\title{Dot and Forest plots data}
\usage{
pre_proc(data, typeval, cod)

prepare_dot_forest_plot_data(
  data,
  drug,
  benefit,
  risk,
  filters,
  category,
  type_graph,
  type_risk,
  ci_method = c("Supplied", "Calculated"),
  space_btwn_out_yn = "Y"
)
}
\arguments{
\item{data}{(\code{data.frame}) dataset.}

\item{typeval}{(\code{character}) Benefit or Risk}

\item{cod}{(\code{character}) identifier either c - continuous or b for binary}

\item{drug}{(\code{character}) selected drug.}

\item{benefit}{(\code{character}) selected benefit.}

\item{risk}{(\code{character}) selected risk.}

\item{filters}{(\code{character}) selected filter.}

\item{category}{(\code{character}) selected category.}

\item{type_graph}{(\code{character}) selected way to display binary outcomes
(absolute risk, relative risk, odds ratio).}

\item{type_risk}{(\code{character}) selected way to display risk outcomes
(crude proportions, exposure-adjusted rates per 100 PYs}

\item{ci_method}{(\code{character}) confidence interval method,
either \code{Supplied} (taken from the effect table)
or \code{Calculated} (calculated within the program).}

\item{space_btwn_out_yn}{(\code{character}) control spacing between outcomes}
}
\description{
Prepare data for Dot and Forest plots
}
\section{Functions}{
\itemize{
\item \code{pre_proc()}: adds a variable in dataframe which
identifies whether the dataframe is for Benefit or Risk analysis.

}}
\examples{
dot_plot_src <- subset(effects_table, !is.na(Prop1))
bdin <- subset(dot_plot_src, Factor == "Benefit")
rdin <- subset(dot_plot_src, Factor == "Risk")

fplot_data <- prepare_dot_forest_plot_data(
  data = dot_plot_src,
  drug = unique(dot_plot_src$Trt1),
  benefit = unique(bdin$Outcome),
  risk = unique(rdin$Outcome),
  filters = "None",
  category = "All",
  type_graph = "Absolute risk",
  type_risk = "Crude proportions",
  ci_method = "Calculated",
  space_btwn_out_yn = "N"
)

}
